Add auto scaling styling to group image and fix navbar items layering on desktop.
On desktop when resizing the browser there is a range of browser sizes that cause
the title and menu options to layer on top of each other, before the menu options
collapse into dropdown menu. This commit fixes this issue on desktop. Also the
group image now auto-resizes up to 33% its original size, and down to the original
200px size to maintain its currently visible size on mobile.